spaces/requirements.txt
marker-pdf==0.2.17
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
